+++ src/spring/azext_spring/tests/latest/test_asa_certificate.py
@@ -5,7 +5,7 @@
 import unittest
 import os
 from .common.test_utils import get_test_cmd
-from ...vendored_sdks.appplatform.v2024_01_01_preview import models
+from ...vendored_sdks.appplatform.v2024_05_01_preview import models
 from ..._utils import _get_sku_name
 from ...custom import (certificate_add, certificate_update)
 try: