+++ src/aks-preview/HISTORY.rst
@@ -13,6 +13,15 @@ Pending
 +++++++
 * support `--yes` for `az aks mesh upgrade rollback` and `az aks mesh upgrade complete` commands.
 
+3.0.0b5
++++++++
+* Add `--bootstrap-artifact-source` and `--bootstrap-container-registry-resource-id` to `az aks update`.
+
+3.0.0b4
++++++++
+* Fix the issue that option `--uptime-sla` is ignored in command `az aks create`.
+* Fix the issue that option `--uptime-sla` and `--no-uptime-sla` are ignored in command `az aks update`.
+
 3.0.0b3
 +++++++
 * Add `--nodepool-initialization-taints` to `az aks create` and `az aks update`.