+++ src/fleet/HISTORY.rst
@@ -90,3 +90,7 @@ Release History
 1.1.1
 ++++++
 * Removed automatic population of empty dns_name_prefix as this is handled server-side now.
+
+1.1.2
+++++++
+* Removed preview markings for hub-related parameters.