+++ src/nginx/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.0.0b2
+++++++
+* use appropriate sku in the example command for deployment
+
 2.0.0b1
 ++++++
 * added autoupgrade and autoscale for nginx deployment. Configuration update added. Flatten Configuration analyze